Feedback from review
* Split up tests for transactions API scenarios.
* Move value parsing into dataset.
