+++ tests/base.py
@@ -234,6 +234,15 @@ class BaseEventsTest(BaseDatasetTest):
         enforce_table_writer(self.dataset).get_writer().write(rows)
 
 
+class BaseApiTest(BaseEventsTest):
+    def setup_method(self, test_method, dataset_name='events'):
+        super().setup_method(test_method, dataset_name)
+        from snuba.api import application
+        assert application.testing is True
+        application.config['PROPAGATE_EXCEPTIONS'] = False
+        self.app = application.test_client()
+
+
 def message(offset, partition, value, eof=False) -> FakeKafkaMessage:
     if eof:
         error = MagicMock()
